Use Rgba<f64> instead of Rgba<u8> and fix a few clippy problems

+++ centaurus_core/src/light/mod.rs
@@ -6,7 +6,7 @@ use std::option::Option;
 #[typetag::deserialize(tag = "type")]
 pub trait Light {
     /// Calculate from which direction and with which amount of color the `position` is hit by the current `Light`.
-    fn hit(&self, position: &Point3<f64>) -> Option<(Vector3<f64>, Rgb<u8>)>;
+    fn hit(&self, position: &Point3<f64>) -> Option<(Vector3<f64>, Rgb<f64>)>;
 }
 
 mod lightbulb;
